makes create functionality work on food and other pages
