+++ src/components/AlternFood.js
@@ -1,4 +1,4 @@
-import React, {useState} from 'react'
+import React, {useState, useEffect} from 'react'
 
 const AlternFood = (props) => {
     const [newForm, setNewForm] = useState({
@@ -7,8 +7,11 @@ const AlternFood = (props) => {
         img: ""
     })
 
+    const [foodAlt, setFoodAlt] = useState(null)
+
+    const altURL = "https://dtox-backend.herokuapp.com/altfood/"
+
 //////// NEEDS TO RECEIVE THE LIST OF ALTERNATIVES FROM THE API AS PROPS PASSED DOWN FROM THE SHOW PAGE.  THEN WE NEED TO MAP OVER THEM AND DISPLAY THEM BELOW THE FORM.//////////////
-    
 
     const handleChange = (e) => {
         setNewForm({...newForm, [e.target.name]: e.target.value})
@@ -16,7 +19,7 @@ const AlternFood = (props) => {
 
     const handleSubmit = (e) => {
         e.preventDefault();
-        ////// call a function that will be defined in the index pages to update state in the index pages to add/post this new alternative to the alternatives database. The index page will also render the list of alternatives.
+        createAlt(newForm)
         setNewForm({
             name: "",
             description: "",
@@ -24,9 +27,33 @@ const AlternFood = (props) => {
         })
     }
 
+    const getAlt = async () => {
+        const response = await fetch(altURL);
+        const data = await response.json();
+        setFoodAlt(data);
+    }
+
+    useEffect(() => {
+        getAlt();
+    }, [])
+
+    ///////////// FUNCTION TO CREATE A NEW ALTERNATIVE //////////////
+    const createAlt = async (alt) => {
+        await fetch(altURL, {
+            method: "POST",
+            headers: {
+                "Content-Type": "Application/json",
+            },
+        
+            body: JSON.stringify(alt)
+        })
+        
+        getAlt()
+      }
+
     //////// Loaded function ///////
     const loaded = () => {
-        return props.foodAlt.map((alt) => {
+        return foodAlt.map((alt) => {
             return (
                 <div className="altDiv">
                     <h3>{alt.name}</h3>
@@ -72,7 +99,7 @@ const AlternFood = (props) => {
                 value="Create Alternative"
             />
         </form>
-        { props.foodAlt ? loaded() : loading()}
+        { foodAlt ? loaded() : loading()}
     </section>
   )
 }
