+++ src/components/Header.js
@@ -5,9 +5,7 @@ import { Link } from 'react-router-dom';
 const Header = () => {
   return (
 
-    <nav className="nav">
-        <Link to="/">
-            <div>HOME</div>
+    
 
     <nav classname="nav">
         <Link id="homepage" to="/">
