+++ src/pages/OtherIndex.js
@@ -5,9 +5,7 @@ import AlternOther from '../components/AlternOther';
 
 const OtherIndex = (props) => {
 
-    const [otherAlt, setOtherAlt] = useState(null)
 
-    const altURL = "https://dtox-backend.herokuapp.com/altother"
  
  /////// LOADING FUNCTION TO DISPLAY JUST THE CHEMICAL NAMES ///////
  const loaded = () => {
@@ -25,15 +23,7 @@ const loading = () => {
   return <h1>Loading . . .</h1>
 }
 
-const getAlt = async () => {
-    const response = await fetch(altURL);
-    const data = await response.json();
-    setOtherAlt(data);
-}
 
-useEffect(() => {
-    getAlt();
-}, [])
 
 
 return (
@@ -42,7 +32,7 @@ return (
       <h3>Select a chemical below to see more!</h3>
       {props.other ? loaded() : loading()}
       <h1>Healthy Alternative Options</h1>
-      <AlternOther otherAlt={otherAlt}/>
+      <AlternOther />
   </div>
 )
 }
